% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBMAT-package.R
\docType{package}
\name{CBMAT-package}
\alias{CBMAT}
\alias{CBMAT-package}
\title{CBMAT: Copula Based Multivariate Association Test for bivariate mixed phenotypes}
\description{
Fit generalised mixed models to test for association between genetic regions and bivariate mixed phenotypes using copulas.
}
\author{
\strong{Maintainer}: Julien St-Pierre \email{stpierre.ju@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
