% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainCBMAT.R
\name{CBMAT}
\alias{CBMAT}
\title{CBMAT main function}
\usage{
CBMAT(
  y1 = NULL,
  fam1 = NULL,
  y2 = NULL,
  fam2 = NULL,
  x = cbind(rep(1, length(y1))),
  G = NULL,
  copfit = c("Gaussian", "Clayton", "Franck", "Gumbell"),
  weight = FALSE,
  weight.para1 = 1,
  weight.para2 = 25,
  pval.method = "min"
)
}
\arguments{
\item{y1}{vector designates the first phenotype (one row-entry per individual), of length \eqn{n}.}

\item{fam1}{a description of the error distribution and link function to be used in the marginal model of the first phenotype. Can be one of the following:
\itemize{
\item \code{"binomial(link=probit)"}
\item \code{"gaussian()"}
\item \code{"Gamma(link=log)"}
\item \code{"Student(link=)"}
}}

\item{y2}{vector designates the second phenotype (one row-entry per individual), of length \eqn{n}.}

\item{fam2}{a description of the error distribution and link function to be used in the marginal model of the second phenotype. Can be one of the following:
\itemize{
\item \code{"gaussian()"}
\item \code{"Gamma(link=log)"}
\item \code{"Student(link=)"}
}}

\item{x}{matrix of covariates including intercept (dimension:\eqn{n \times k}, with \eqn{k} the number of covariates), Default: cbind(rep(1, length(y1)))}

\item{G}{matrix of SNPs (dimension:\eqn{n \times p}, with \eqn{p} the number of SNPs)}

\item{copfit}{character, selects the copula(s) to use for modelling phenotypes dependence.
Can be any one of the following:
\itemize{
\item \code{"Gaussian"}, Gaussian copula
\item \code{"Clayton"}, Clayton copula
\item \code{"Gumbel"}, Gumbel copula
\item \code{"Frank"}, Frank copula
}, Default: c("Gaussian", "Clayton", "Franck", "Gumbell")}

\item{weight}{logical, should weights be used to increase power for rare variants, Default: FALSE}

\item{weight.para1}{alpha parameter of beta distribution used to simulate weights, Default: 1}

\item{weight.para2}{beta parameter of beta distribution used to simulate weights, Default: 25}

\item{pval.method}{character, which method to use to calculate p-value.
Can be one of the following:
\itemize{
\item \code{"min"} (default)
\item \code{"Fischer"}, Fischer
\item \code{"MFKM"}, MFKM
}}
}
\value{
A list containing results of the association test
specified by the \code{method} parameter using the copula model for modelling primary-secondary dependency
by the \code{copfit} parameter for the sampling mechanism \code{Design}. The
output list contains the following results:
\itemize{
\item \code{p.value}: the p-value of the region-based association test
\item \code{gamma.y1}: the intercept and covariates estimates from the marginal model of the first phenotype
\item \code{gamma.y2}: the intercept and covariates estimates from the marginal model of the second phenotype
\item \code{alpha}: the estimated phenotypic dependence parameter, as a parameter of the used copula
\item \code{tau}: Kendall's tau, which is a function of the estimated phenotypic dependence parameter alpha
\item \code{cop}: based on the AIC criterion calculated based on the copula-model
}
}
\description{
FUNCTION_DESCRIPTION
}
\details{
When "weight=TRUE", a weighted test is used. The weight for each SNP is a beta fuction of the corresponding minor allele frequency (MAF).
There are two parameters, weight.para1 and weight.para2, for beta function. For example, when weight.para1=weight.para2=0.5, the corresponding weight is 1/sqrt(MAF*(1-MAF));
when weight.para=1 and weight.para=25, the corresponding weight is the one suggested by SKAT.
}
\author{
Julien St-Pierre and Karim Oualkacha
}
